Move `fontique::GenericFamily` to `styled_text`
